Revert "wip software restructure"
This reverts commit 9efdc30148ed82dc11bf6d7a88d7bcb034e4f574.
