--- FlexSensor.hpp
@@ -1,12 +1,0 @@
-#ifndef TENACI_FORCE_SENSOR_H
-#define TENACI_FORCE_SENSOR_H
-
-#include "Sensor.hpp"
-
-class ForceSensor : public Sensor
-{
-
-};
-
-
-#endif // TENACI_FORCE_SENSOR_H
