--- ForceSensor.hpp
@@ -1,7 +1,0 @@
-#ifndef TENACI_FORCE_SENSOR_H
-#define TENACI_FORCE_SENSOR_H
-
-
-
-
-#endif // TENACI_FORCE_SENSOR_H
