--- SensitivitySensor.hpp
@@ -1,6 +1,0 @@
-#ifndef TENACI_SENSITIVITY_SENSOR_H
-#define TENACI_SENSITIVITY_SENSOR_H
-
-
-
-#endif // TENACI_SENSITIVITY_SENSOR_H
